Backend/requirements.txt
Flask
python-dotenv
supabase-py
flask-cors
